Even more solvers moved, trying to resolve test conflicts
